# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Example API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cats/{cat}": {
            "get": {
                "tags": [
                    "cats"
                ],
                "summary": "Get Cat",
                "description": "Get an ascii cat",
                "operationId": "600644e35af3223e90be7f419c3f266a",
                "parameters": [
                    {
                        "name": "cat",
                        "in": "path",
                        "description": "Cat name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "achilles",
                                "hercules"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Unknown cat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The selected cat is invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "{'errors':{'cat':['The selected cat is invalid.']}}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}